Improve `on-complete` functions
Providing a 2-args arity make those functions useful besides being
`on-complete` functions.

Sometimes, it can be that steps are skipped for some reasons (eg. a live
animation following running milliseconds). Hence, a transition might
jump past 100% and the final state would not be computed. On the other
hand, this is often unacceptable. Using `fn-assoc-data` is a simple way
to ensure the final value - which is typically known in advance - is
indeed computed.