Add propagation of state in poly-transitions
A poly-transition is created by providing a list of functions returning
transitions. Those functions now receive an additional parameter which
is the state at the moment the transitions are created, which is hugely
important for doing poly-transition because each transition most likely
depends on the state reached by the previous one.