Prepare for 1.0.0 release
Full blown API, purely discrete event driven, also acting as a backend
for simulating (at leat some) continuous and hybrid phenomena.

The new design naturally allows for easily building complex activities
in the DES sense, any sequences of continuous flows intertwined or not
with regular events, as well as any form of useful periodicity. This is
done by treating everything as an event, even the sampling of flows, and
events can be composed using queues in an interceptor-like methodology
instead of traditional functional composition.