Clarify error handling and add canceling flat event
An ::on-error function attached to a queue as metada receives a map
containing the ctx as it was before executing the function of the queue
that led to the exception, as well as the state of the queue. Both are
kept separated, that queue being removed from the context, because it
would be all to easy to forget about it and inadvertently and re-execute
it.

If the exception is handled at an outer queue, the map also contains the
inner context and its inner queue, so that we know the state of
everything right before the exception happened.