Fix interruption of nested queues + finalize error handling
The way it was implemented, it was not actually possible to cancel outer
queues. In nested queue handling, named recursion has been replaced with
tail call recursion and a virtualized stack kept at ::e-flat, thus easy
to remove.

As it turns out, having this stack clearly simplified exception handling
at the level of queues.