Bring flow sampling deduplication without requests
By keeping track of which flows have been sampled at a given ptime by
dissociating them from a copy of the flow tree, sampling is now executed
directly at the level of events instead as a separate further step. Not
only is it more convenient, it follows the idiom of treating samples as
events.

Tests involving flows and operations do not pass. However, the
implementation of operations is lacking and needs to be improved.