Remove operations and leverage the `dvlopt.fdat` lib
FDat is a library that has been created specifically for serializing
function and not having to use any "event vectors" or those "operations"
we were implementing.

Now, functions that are meant to appear in queues such as for creating
flows or inducing a delay are simply annotated using the `?` macro from
FDat. This is both minimalistic and efficient.

[https://github.com/dvlopt/fdat.cljc](https://github.com/dvlopt/fdat.cljc)